requirements.txt
aiogram==3.10.0
aiomysql==0.2.0
python-dotenv==1.0.1
openai==1.40.0
aiohttp==3.10.5
